重構音頻和資源管理代碼，統一使用unsigned char和unsigned short類型，提升代碼一致性和可讀性；優化YJ2解壓縮邏輯

+++ src/global.c
@@ -696,7 +696,7 @@ int PAL_AddItemToInventory(
     }
 }
 
-int PAL_GetItemAmount(
+unsigned short PAL_GetItemAmount(
     unsigned short wItem)
 /*++
   Purpose:
